refactor(edge): isolate quic listener connection registry ownership
